Splinter customer quote logic into independent component
